pyproject.toml
[tool.poetry]
name = "myobservatory-forecast-check"
version = "0.1.0"
description = ""
authors = ["Simon Poon <kalunpoon926@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.13"
appium-python-client = "5.0.0"
selenium = "4.31.0"
pytest-html = "4.1.1"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
